requirements.txt
PyPDF2
langdetect
deep-translator
google-generativeai
pymongo
flask
google-cloud-aiplatform
google-auth-oauthlib
